% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_likelihood.R
\name{cospec_cond_loglik}
\alias{cospec_cond_loglik}
\title{Conditional censored likelihood}
\usage{
cospec_cond_loglik(x, censoring, mu, tau)
}
\arguments{
\item{x}{[numeric] vector of observations}

\item{censoring}{[logical] vector; if \code{TRUE}, observation is left-censored, observed otherwise}

\item{mu}{[numeric] mean vector}

\item{tau}{[numeric] standard deviation vector (or scalar)}
}
\description{
Censored likelihood for observations from the conditional
spatial extremes model, given the imputed random effects
Each spatio-temporal observation is conditionally independent
of one another, so the likelihood is formed from the
product of univariate normal density and distribution functions
}
