% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_likelihood.R
\name{cospec_cond_loglik_grad_tau}
\alias{cospec_cond_loglik_grad_tau}
\title{Gradient of conditional censored log likelihood}
\usage{
cospec_cond_loglik_grad_tau(x, censoring, mu, tau)
}
\arguments{
\item{x}{[numeric] vector of observations}

\item{censoring}{[logical] if \code{TRUE}, observation is left-censored, observed otherwise}

\item{mu}{[numeric] mean vector}

\item{tau}{[numeric] standard deviation vector (or scalar)}
}
\description{
Gradient with respect to scale \eqn{\tau}.
}
