% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_likelihood.R
\name{cospec_cond_loglik_hessian_mu}
\alias{cospec_cond_loglik_hessian_mu}
\title{Hessian of conditional censored log likelihood}
\usage{
cospec_cond_loglik_hessian_mu(x, censoring, mu, tau)
}
\arguments{
\item{x}{[numeric] vector of observations}

\item{censoring}{[logical] vector; if \code{TRUE}, observation is left-censored, observed otherwise}

\item{mu}{[numeric] mean vector}

\item{tau}{[numeric] standard deviation vector (or scalar)}
}
\value{
a list with two vectors, to be used with the chain rule
}
\description{
Hessian with respect to location parameter.
}
