% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical_slice_sampling.R
\name{elliptical_slice_sampling}
\alias{elliptical_slice_sampling}
\title{Elliptical slice sampling}
\usage{
elliptical_slice_sampling(curState, logLikFn, normSamp, ...)
}
\arguments{
\item{curState}{current value of the random state vector}

\item{logLikFn}{log likelihood function for the model, whose first arguemnt is \code{x}, corresponding to \code{curState}}

\item{normSamp}{Gaussian random vector}

\item{...}{additional arguments passed to \code{logLikFn}}
}
\value{
a list with elements
\itemize{
\item{x}{vector of \code{length(curState)}}
\item{niter}{number of iterations}
}
}
\description{
This function implements the elliptical slice sampler of
Murray, Adams and MacKay (2012) given a random sample from a Gaussian distribution with mean zero and covariance matrix. Sampling is performed outside of the function to allow users to leverage existing functions for sparse matrices.
}
