% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{ScalingFunction}
\alias{ScalingFunction}
\title{Affine renormalization function}
\usage{
ScalingFunction(
  npar,
  default,
  func,
  grad,
  type = "scale",
  lower = rep(-Inf, npar),
  upper = rep(Inf, npar),
  cst = TRUE,
  fixed = rep(FALSE, npar)
)
}
\arguments{
\item{npar}{number of parameters passed as first argument to \code{func} and \code{grad}}

\item{func}{function of \code{npar} argument and distance}

\item{grad}{gradient of \code{func} with respect to parameters}

\item{type}{integer indicating the type of function, either \code{scale}, \code{loc} or \code{locx} for locations (\code{locx}) gets multiplied by \code{x}, the value at the conditioning site}

\item{lower}{\code{npar} vector of lower bounds for the parameter vector}

\item{upper}{\code{npar} vector of upper bounds for the parameter vector}

\item{cst}{logical; is the function constant. If \code{FALSE}, this indicates that \code{func} is a two-place function that returns}

\item{fixed}{\code{npar} logical vector indicating whether parameters are held fixed to their \code{default} values}
}
\description{
This class is used to define a location or scaling function.
}
