% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_normal.R
\name{dtnorm}
\alias{dtnorm}
\title{Density of a univariate truncated Normal}
\usage{
dtnorm(x, mean = 0, sd = 1, a = -Inf, b = Inf, log = FALSE)
}
\arguments{
\item{x}{[numeric] observation vector}

\item{mean}{[numeric] scalar location parameter}

\item{sd}{[numeric] scalar scale parameter}

\item{a}{[numeric] lower bound}

\item{b}{[numeric] upper bound}

\item{log}{[logical] if \code{TRUE}, return the log density}
}
\description{
Density of a univariate truncated Normal
}
